Fix: example bug

+++ examples/streams-i2s-serial/streams-i2s-serial.ino
@@ -1,18 +1,17 @@
-/**
- * @file streams-i2s-serial_16bit.ino
+ /**
+ * @file streams-i2s-serial.ino
  * @author Phil Schatzmann
- * @brief see https://github.com/pschatzmann/arduino-audio-tools/blob/main/examples/examples-stream/streams-i2s-serial_16bit/README.md
+ * @brief see https://github.com/pschatzmann/arduino-audio-tools/blob/main/examples/examples-stream/streams-i2s-serial/README.md
  * 
  * @author Phil Schatzmann
  * @copyright GPLv3
  */
 
-
 #include "AudioTools.h"
 
 AudioInfo info(16000, 2, 32);
 I2SStream i2sStream; // Access I2S as stream
-CsvOutput<int16_t> csvStream(Serial);
+CsvOutput<int32_t> csvStream(Serial);
 StreamCopy copier(csvStream, i2sStream); // copy i2sStream to csvStream
 
 // Arduino Setup
@@ -23,10 +22,8 @@ void setup(void) {
     auto cfg = i2sStream.defaultConfig(RX_MODE);
     cfg.copyFrom(info);
     cfg.i2s_format = I2S_STD_FORMAT; // or try with I2S_LSB_FORMAT
-    cfg.is_master = true;
-     // this module nees a master clock if the ESP32 is master
-    cfg.use_apll = false;  // try with yes
-    //cfg.pin_mck = 3; 
+    cfg.is_master = true;  // this module nees a master clock if the ESP32 is master
+    cfg.use_apll = false;   // try with yes
     i2sStream.begin(cfg);
 
     // make sure that we have the correct channels set up
